requirements.txt
qrcode==6.1
Django==3.1.1
django_filter==2.4.0
Pillow==8.0.1
django-environ==0.4.5
